Viseur final
Le viseur est visible et fonctionnel (la texture est multipliée pour qu'on le voie bien).
Par contre, y'a un problème car quand on bouge le curseur sur l'écran du jeu, la souris bouge derrière sur le bureau, et n'a pas la même position que le curseur.

+++ Source/HUD/Cursor.h
@@ -7,7 +7,7 @@ public:
 	~Cursor();
 
 	sf::Vector2i position;
-	sf::Vector2i getPosition();
+	sf::Vector2f getPosition();
 	void update();
 	void draw(sf::RenderTarget& target, sf::RenderStates states) const;
